Refactor CondaSearch code; make search more robust to errors; remove duplicate chips; search on ENTER
